Fix dynamic linking error
- Improve image usage by adding the scripts to the path

After building the executables on an debian based image it expects
some libs for the dynamic linking on some locations that do not exist
on alpine linux. I tried to change the target to a musl linx but it
wasn't available for the newest nightly rust build. Now the final image
is based on debian. This increases the image size to ca. 100MB.