Set fix version commit hash for stable builds
Set commit hash to get a working version of all binaries.
Improves build time by splitting the build of the rust binaries and the
caching for the ui building.
At this point in time there is no good way of caching rust
dependencies. This is why it has to download and compile everything on a
change.